fix subdock positioning
Subdocks are now seem to be positioned properly. Also, pointing arrows work
"magically" (I don't understand why, I assumed they would point to wrong
locations).

Does not work: moving an applet to a different dock (subdock positioning is
messed up after and needs Cairo-Dock to be restarted); I'm guessing this
should be fixable by setting the parent to the new dock at the right location.

Not tested: vertical docks

Additional problem: this would normally need GDK 3.24 at least. I've added
a hack (based on code in gtk-layer-shell) to make it work on GDK 3.22, by using
an unexposed internal function, since I did not want to upgrade my GDK
libraries (this is in the previous commit).
Probably this should be removed in the future.