add cmake option to use new positioning code on X11 as well
New code for positioning menus and subdocks should work on both Wayland and X11.
Now, by default, it is only used on Wayland, and the old code is used on X11 to
avoid regressions. This option can be used to test the new code on X11 as well;
at a later point, it could be made the default to simplify things.