Fix for not properly handling the pointer leaving the dock on Wayland.
Since we do not have global mouse positions on Wayland, determining if the mouse is inside the dock
depends entirely on the enter / leave events sent by GDK. Some of the renderers might still rely on
mouse position which might cause a problem. I've added changes to hopefully avoid these (as far as
I could test so far).

Also handle subdocks being closed on Wayland

On Wayland, the compositor can choose to close (unmap) subdocks at any time,
since they are xdg-popups. This typically happens when an app is activated
(e.g. by clicking on an icon). This results in the main dock being in a
"frozen" state, since it is not notified that it does not have the focus
anymore.

This change handles the unmap events on docks and invokes the pointer leave
handler to take care of this.