Allow keyboard interactivity
This is required for the dock to receive any keyboard events.

Requires support for layer-shell version >= 4 in the compositor and in gtk-layer-shell.

This will compile cairo-dock with keyboard interactivity enabled if the gtk-layer-shell library at compile time supports it. TODO: make this optional, so that cairo-dock can be compiled with a newer version of gtk-layer-shell installed, but used with an older version (or just require gtk-layer-shell 0.5.2 or newer for this).